Adopt new auth token scheme for sdmGet and sdmPut.
With the addition of sdmAuth users will now have to authenticate via
google prior to being able to use sdmGet and sdmPut. In addition they
will be required to configure a given instance (e.g.,
sni-sdm.stanford.edu) by entering the client_secret/id for that instance -
which they will be prompted for the first time they run sdmGet or put.